requirements.txt
# Core API Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# AI Model APIs
google-generativeai==0.3.1
openai==1.3.5
anthropic==0.7.1

# Document Processing (Simplified - no pandas/numpy dependencies)
PyPDF2==3.0.1
python-docx==1.1.0
openpyxl==3.1.2

# Document Generation
reportlab==4.0.7
python-docx==1.1.0

# OCR and Image Processing
Pillow==10.1.0

# Code Execution (Sandboxed)
RestrictedPython==6.2

# Memory System
aiosqlite==0.19.0

# Utilities
python-dotenv==1.0.0
httpx==0.25.1
aiofiles==23.2.1
requests==2.31.0
